move config to individual package
