+++ go.mod
@@ -4,18 +4,18 @@ go 1.12
 
 require (
 	github.com/amaretto/waveform v0.0.0-20210512181107-2372539edf07
-	github.com/benjojo/bpm v0.0.0-20181021194517-7b9b0c63ec95 // indirect
-	github.com/dhowden/tag v0.0.0-20201120070457-d52dcb253c63 // indirect
+	github.com/benjojo/bpm v0.0.0-20181021194517-7b9b0c63ec95
+	github.com/dhowden/tag v0.0.0-20201120070457-d52dcb253c63
 	github.com/faiface/beep v1.0.2
 	github.com/gdamore/tcell v1.4.0
 	github.com/gdamore/tcell/v2 v2.1.0
-	github.com/golang/protobuf v1.4.2 // indirect
+	github.com/golang/protobuf v1.4.2
 	github.com/hajimehoshi/go-mp3 v0.3.1 // indirect
 	github.com/mattn/go-sqlite3 v2.0.3+incompatible
 	github.com/rivo/tview v0.0.0-20210117162420-745e4ceeb711
-	github.com/sirupsen/logrus v1.8.1 // indirect
+	github.com/sirupsen/logrus v1.8.1
 	github.com/spf13/cobra v1.1.1
 	gonum.org/v1/plot v0.8.1 // indirect
-	google.golang.org/grpc v1.38.0 // indirect
-	gopkg.in/yaml.v2 v2.4.0 // indirect
+	google.golang.org/grpc v1.38.0
+	gopkg.in/yaml.v2 v2.4.0
 )
