+++ pkg/cmd/cli/config/config_test.go
@@ -1,4 +1,4 @@
-package player
+package config
 
 import (
 	"os"
